requirements.txt
aiohttp==3.8.5
aioice==0.9.0
aiortc==1.5.0
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
av==10.0.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cryptography==41.0.3
dnspython==2.4.2
fastapi==0.103.1
frozenlist==1.4.0
google-crc32c==1.5.0
h11==0.14.0
idna==3.4
ifaddr==0.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
numpy==1.25.2
opencv-python==4.8.0.76
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pyee==11.0.0
pylibsrtp==0.8.0
pyOpenSSL==23.2.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.2
websockets==11.0.3
yarl==1.9.2
onnxruntime==1.14.1
opencv-contrib-python==4.7.0.72
claude_api==1.0.17
python-dotenv== 1.0.0